Compiles cleanly.

+++ internal/ban/report.go
@@ -76,7 +76,7 @@ func (status ReportStatus) String() string {
 }
 
 type Report struct {
-	ReportID        int64           `json:"report_id"`
+	ReportID        int32           `json:"report_id"`
 	SourceID        steamid.SteamID `json:"source_id"`
 	TargetID        steamid.SteamID `json:"target_id"`
 	Description     string          `json:"description"`
@@ -117,7 +117,7 @@ type ReportWithAuthor struct {
 }
 
 type ReportMessage struct {
-	ReportID        int64                `json:"report_id"`
+	ReportID        int32                `json:"report_id"`
 	ReportMessageID int64                `json:"report_message_id"`
 	AuthorID        steamid.SteamID      `json:"author_id"`
 	MessageMD       string               `json:"message_md"`
@@ -133,7 +133,7 @@ func (m ReportMessage) Path() string {
 	return fmt.Sprintf("/report/%d#%d", m.ReportID, m.ReportMessageID)
 }
 
-func NewReportMessage(reportID int64, authorID steamid.SteamID, messageMD string) ReportMessage {
+func NewReportMessage(reportID int32, authorID steamid.SteamID, messageMD string) ReportMessage {
 	now := time.Now()
 
 	return ReportMessage{
@@ -249,7 +249,7 @@ func (r Reports) addAuthorsToReports(ctx context.Context, reports []Report) ([]R
 	return userReports, nil
 }
 
-func (r Reports) SetReportStatus(ctx context.Context, reportID int64, user personDomain.BaseUser, status ReportStatus) (ReportWithAuthor, error) {
+func (r Reports) SetReportStatus(ctx context.Context, reportID int32, user personDomain.BaseUser, status ReportStatus) (ReportWithAuthor, error) {
 	report, errGet := r.Report(ctx, user, reportID)
 	if errGet != nil {
 		return report, errGet
@@ -287,7 +287,7 @@ func (r Reports) SetReportStatus(ctx context.Context, reportID int64, user perso
 	))
 
 	slog.Info("Report status changed",
-		slog.Int64("report_id", report.ReportID),
+		slog.Int64("report_id", int64(report.ReportID)),
 		slog.String("to_status", report.ReportStatus.String()))
 
 	return report, nil
@@ -323,7 +323,7 @@ func (r Reports) Reports(ctx context.Context) ([]ReportWithAuthor, error) {
 	return r.addAuthorsToReports(ctx, reports)
 }
 
-func (r Reports) Report(ctx context.Context, curUser personDomain.BaseUser, reportID int64) (ReportWithAuthor, error) {
+func (r Reports) Report(ctx context.Context, curUser personDomain.BaseUser, reportID int32) (ReportWithAuthor, error) {
 	report, err := r.repository.GetReport(ctx, reportID)
 	if err != nil {
 		return ReportWithAuthor{}, err
@@ -346,7 +346,7 @@ func (r Reports) Report(ctx context.Context, curUser personDomain.BaseUser, repo
 	var demo servers.DemoFile
 	if report.DemoID > 0 {
 		if errDemo := r.demos.GetDemoByID(ctx, report.DemoID, &demo); errDemo != nil {
-			slog.Error("Failed to load report demo", slog.Int64("report_id", report.ReportID))
+			slog.Error("Failed to load report demo", slog.Int64("report_id", int64(report.ReportID)))
 		}
 	}
 
@@ -362,7 +362,7 @@ func (r Reports) ReportBySteamID(ctx context.Context, authorID steamid.SteamID, 
 	return r.repository.GetReportBySteamID(ctx, authorID, steamID)
 }
 
-func (r Reports) Messages(ctx context.Context, reportID int64) ([]ReportMessage, error) {
+func (r Reports) Messages(ctx context.Context, reportID int32) ([]ReportMessage, error) {
 	return r.repository.GetReportMessages(ctx, reportID)
 }
 
@@ -466,7 +466,7 @@ func (r Reports) Save(ctx context.Context, currentUser personDomain.BaseUser, re
 		return ReportWithAuthor{}, err
 	}
 
-	slog.Info("New report created", slog.Int64("report_id", report.ReportID))
+	slog.Info("New report created", slog.Int64("report_id", int64(report.ReportID)))
 
 	if demo.DemoID > 0 && !demo.Archive {
 		if errMark := r.demos.MarkArchived(ctx, &demo); errMark != nil {
@@ -530,7 +530,7 @@ func (r Reports) EditMessage(ctx context.Context, reportMessageID int64, curUser
 	return r.MessageByID(ctx, reportMessageID)
 }
 
-func (r Reports) CreateMessage(ctx context.Context, reportID int64, curUser personDomain.BaseUser, req RequestMessageBodyMD) (ReportMessage, error) {
+func (r Reports) CreateMessage(ctx context.Context, reportID int32, curUser personDomain.BaseUser, req RequestMessageBodyMD) (ReportMessage, error) {
 	req.BodyMD = strings.TrimSpace(req.BodyMD)
 
 	if req.BodyMD == "" {
@@ -574,7 +574,7 @@ func (r Reports) CreateMessage(ctx context.Context, reportID int64, curUser pers
 
 	sid := curUser.GetSteamID()
 	slog.Info("New report message created",
-		slog.Int64("report_id", reportID), slog.String("steam_id", sid.String()))
+		slog.Int64("report_id", int64(reportID)), slog.String("steam_id", sid.String()))
 
 	return msg, nil
 }
